add mshadow binary io

+++ cxxnet/cxxnet_main.cpp
@@ -240,8 +240,9 @@ namespace cxxnet{
                 net_trainer->Predict(pred, batch);
                 utils::Assert( batch.num_batch_padd < batch.batch_size, "num batch pad must be smaller");
                 pred.resize( pred.size() - batch.num_batch_padd );
+
                 for (mshadow::index_t j = 0; j < pred.size(); ++j) {
-                    fprintf(fo, "%d\n", static_cast<int>(pred[j]));
+                    fprintf(fo, "%g\n", pred[j]);
                 }
             }
             fclose( fo );
